feat(chat): refactor ChatWidget component and add CSS styles
